added multithreading and redis cache

+++ 16_CorrespondenceApi/src/main/java/com/rihs/binding/CorrespondenceResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CorrespondenceResponse {
 
-	private Integer totalTriggers;
-	private Integer successTriggers;
-	private Integer failedTriggers;
+	private Long totalTriggers;
+	private Long successTriggers;
+	private Long failedTriggers;
 }
